refactor: extract mini navi overlay
